Move access token retrival server side
