+++ packages/office-addin-sso/.gitignore
@@ -1,0 +1,5 @@
+# Build output
+lib/
+
+# Node modules folder
+node_modules/
