+++ packages/office-addin-sso/package.json
@@ -41,7 +41,7 @@
     "@types/cors": "2.8.4",
     "@types/express": "4.17.13",
     "@types/jquery": "^3.3.34",
-    "@types/mocha": "^5.2.7",
+    "@types/mocha": "^9.1.1",
     "@types/morgan": "1.7.35",
     "concurrently": "^6.2.2",
     "copyfiles": "2.4.1",
