+++ packages/office-addin-sso/src/middle-tier/odata-helper.ts
