+++ packages/office-addin-sso/tsconfig.json
@@ -11,12 +11,6 @@
         "sourceMap": true,
         "strict": false,
         "target": "es6",
-        "types": [
-            "node"
-        ],
-        "typeRoots": [
-            "./node_modules/@types/"
-        ]
     },
     "include": [
         "src/**/*",
